fix bug

+++ db/activity-3.0.0-delta-schema.sql
@@ -11,7 +11,7 @@ ALTER TABLE `eh_activity_roster` ADD COLUMN `refund_time` DATETIME NULL;
 ALTER TABLE `eh_activity_roster` ADD COLUMN `status` TINYINT(4) DEFAULT '2' NULL COMMENT '0: cancel, 1: reject, 2:normal';
 
 ALTER TABLE `eh_activities` ADD COLUMN `charge_flag` TINYINT(4) DEFAULT '0' NULL COMMENT '0: no charge, 1: charge'; 
-ALTER TABLE `eh_activities` ADD COLUMN `charge_price` INT(11) NULL COMMENT 'charge_price';
+ALTER TABLE `eh_activities` ADD COLUMN `charge_price` DOUBLE NULL COMMENT 'charge_price';
 
 
 -- 订单过期时间的设置表
